Add validations

+++ actors-backend/routers/awards.js
@@ -1,4 +1,4 @@
-import {verifyToken} from "../utils/authentication.js";
+import {verifyAdmin, verifyToken} from "../utils/authentication.js";
 import {createAward, deleteAward, getAward, getAwards, modifyAward} from "../controllers/awards.js";
 
 export const awardsRouter = async (req, res) => {
@@ -19,16 +19,19 @@ export const awardsRouter = async (req, res) => {
         req.params.id = id;
         console.log(`DELETE /awards/${id}`);
         verifyToken(req);
+        verifyAdmin(req);
         await deleteAward(req, res);
     } else if (req.fullUrl.pathname === '/' && req.method === 'POST') {
         console.log("POST /awards");
         verifyToken(req);
+        verifyAdmin(req);
         await createAward(req, res);
     } else if (regexRouteWithId.test(req.fullUrl.pathname) && req.method === 'PUT') {
         let id = parseInt(req.fullUrl.pathname.match(regexRouteWithId)[1], 10);
         req.params.id = id;
         console.log(`PUT /awards/${id}`);
         verifyToken(req);
+        verifyAdmin(req);
         await modifyAward(req, res);
     } else {
         req.handled = false;
